# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cms/special/list": {
      "get": {
        "operationId": "SpecialController_findList",
        "summary": "获取专题列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "title",
            "required": false,
            "in": "query",
            "description": "专题标题",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0禁用 1启用",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/SpecialEntity"
                              }
                            },
                            "total": {
                              "type": "number",
                              "default": 0
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-专题管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/special/{id}": {
      "get": {
        "operationId": "SpecialController_findOne",
        "summary": "获取专题详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SpecialEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-专题管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SpecialController_delete",
        "summary": "删除专题",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-专题管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/special": {
      "post": {
        "operationId": "SpecialController_create",
        "summary": "新增专题",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpecialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-专题管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "SpecialController_update",
        "summary": "修改专题",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpecialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-专题管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/link/list": {
      "get": {
        "operationId": "LinkController_findList",
        "summary": "获取友情链接列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "linkName",
            "required": false,
            "in": "query",
            "description": "链接名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0禁用 1启用",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/LinkEntity"
                              }
                            },
                            "total": {
                              "type": "number",
                              "default": 0
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-友情链接管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/link/{id}": {
      "get": {
        "operationId": "LinkController_findOne",
        "summary": "获取友情链接详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/LinkEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-友情链接管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "LinkController_delete",
        "summary": "删除友情链接",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-友情链接管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/cms/quick-link/list": {
      "get": {
        "operationId": "QuickLinkController_findList",
        "summary": "获取快捷链接列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "linkName",
            "required": false,
            "in": "query",
            "description": "链接名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0禁用 1启用",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/QuickLinkEntity"
                              }
                            },
                            "total": {
                              "type": "number",
                              "default": 0
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-快捷链接管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/quick-link/{id}": {
      "get": {
        "operationId": "QuickLinkController_findOne",
        "summary": "获取快捷链接详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QuickLinkEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-快捷链接管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "QuickLinkController_delete",
        "summary": "删除快捷链接",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-快捷链接管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/quick-link": {
      "post": {
        "operationId": "QuickLinkController_create",
        "summary": "新增快捷链接",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuickLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-快捷链接管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "QuickLinkController_update",
        "summary": "修改快捷链接",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuickLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-快捷链接管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "ResultData": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "default": 200
          },
          "msg": {
            "type": "string",
            "default": "操作成功"
          }
        },
        "required": [
          "code",
          "msg"
        ]
      },
...
      "SpecialEntity": {
        "type": "object",
        "properties": {}
      },
      "CreateSpecialDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "专题标题"
          },
          "description": {
            "type": "string",
            "description": "专题描述"
          },
          "coverImage": {
            "type": "string",
            "description": "封面图"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序",
            "default": 0
          },
          "status": {
            "type": "string",
            "description": "状态：0禁用 1启用",
            "default": "1"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateSpecialDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "专题标题"
          },
          "description": {
            "type": "string",
            "description": "专题描述"
          },
          "coverImage": {
            "type": "string",
            "description": "封面图"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序",
            "default": 0
          },
          "status": {
            "type": "string",
            "description": "状态：0禁用 1启用",
            "default": "1"
          },
          "specialId": {
            "type": "number",
            "description": "专题ID"
          }
        },
        "required": [
          "title",
          "specialId"
        ]
      },
      "LinkEntity": {
        "type": "object",
        "properties": {}
      },
...
      "QuickLinkEntity": {
        "type": "object",
        "properties": {}
      },
      "CreateQuickLinkDto": {
        "type": "object",
        "properties": {
          "linkName": {
            "type": "string",
            "description": "链接名称"
          },
          "linkUrl": {
            "type": "string",
            "description": "链接地址"
          },
          "linkPosition": {
            "type": "string",
            "description": "链接位置：top顶部 sidebar侧边栏",
            "default": "top"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序",
            "default": 0
          },
          "status": {
            "type": "string",
            "description": "状态：0禁用 1启用",
            "default": "1"
          }
        },
        "required": [
          "linkName",
          "linkUrl"
        ]
      },
      "UpdateQuickLinkDto": {
        "type": "object",
        "properties": {
          "linkName": {
            "type": "string",
            "description": "链接名称"
          },
          "linkUrl": {
            "type": "string",
            "description": "链接地址"
          },
          "linkPosition": {
            "type": "string",
            "description": "链接位置：top顶部 sidebar侧边栏",
            "default": "top"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序",
            "default": 0
          },
          "status": {
            "type": "string",
            "description": "状态：0禁用 1启用",
            "default": "1"
          },
          "linkId": {
            "type": "number",
            "description": "链接ID"
          }
        },
        "required": [
          "linkName",
          "linkUrl",
          "linkId"
        ]
      }
    }
  }
}